devtools/client/shared/components/test/mochitest/test_searchbox-with-autocomplete.html

<!-- This Source Code Form is subject to the terms of the Mozilla Public
   - License, v. 2.0. If a copy of the MPL was not distributed with this
   - file, You can obtain one at http://mozilla.org/MPL/2.0/. -->
<!DOCTYPE html>
<html>
<!--
Test the searchbox and autocomplete-popup components
-->
<head>
  <meta charset="utf-8">
  <title>SearchBox component test</title>
  <script src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
  <script src="chrome://mochikit/content/tests/SimpleTest/SpawnTask.js"></script>
  <script src="chrome://mochikit/content/tests/SimpleTest/EventUtils.js"></script>
  <link rel="stylesheet" type="text/css" href="chrome://mochikit/content/tests/SimpleTest/test.css">
</head>
<body>
<script src="head.js"></script>
<script>
/* import-globals-from head.js */
"use strict";
window.onload = async function () {
  /**
   * Takes a DOMNode with its children as list items,
   * Typically UL > LI and each item's text value is
   * compared with the reference item's value as a test
   *
   * @params {Node} - Node to be compared
   * @reference {array} - Reference array for comparison
   */
  function compareAutocompleteList(list, reference) {
    let items = [...list.children].map(el => el.textContent);
    for (let i = 0; i < items.length; i++) {
      let item = items[i];
      let ref = reference[i];
      is(item, ref, `Item ${i} in list is correct`);
    }
  }

  let React = browserRequire("devtools/client/shared/vendor/react");
  let SearchBox = React.createFactory(
    browserRequire("devtools/client/shared/components/search-box")
  );
  const { component, $ } = await createComponentTest(SearchBox, {
    type: "search",
    autocompleteProvider: (filter) => {
      let baseList = [
        "foo",
        "BAR",
        "baZ",
        "abc",
        "pqr",
        "xyz",
        "ABC",
        "a1",
        "a2",
        "a3",
        "a4",
        "a5",
      ];
      if (!filter) {
        return [];
      }

      let tokens = filter.split(/\s+/g);
      let lastToken = tokens[tokens.length - 1];
      let previousTokens = tokens.slice(0, tokens.length - 1);

      if (!lastToken) {
        return [];
      }

      return baseList
        .filter((item) => {
          return item.toLowerCase().startsWith(lastToken.toLowerCase())
            && item.toLowerCase() !== lastToken.toLowerCase();
        })
        .sort()
        .map(item => ({
          value: [...previousTokens, item].join(" "),
          displayValue: item,
        }));
    },
    onChange: () => null,
  });
  const { refs } = component;

  async function testSearchBoxWithAutocomplete() {
    ok(!$(".devtools-autocomplete-popup"), "Autocomplete list not visible");

    $(".devtools-searchinput").focus();
    await forceRender(component); // Wait for state update
    ok(!$(".devtools-autocomplete-popup"), "Autocomplete list not visible");

    sendString("a");
    await forceRender(component);

    compareAutocompleteList($(".devtools-autocomplete-listbox"), [
      "ABC",
      "a1",
      "a2",
      "a3",
      "a4",
      "a5",
      "abc",
    ]);

    is(refs.autocomplete.state.selectedIndex, -1, "Initialised selectedIndex is -1");

    // Blur event
    $(".devtools-searchinput").blur();
    await forceRender(component);
    ok(!component.state.focused, "focused state was properly set");
    ok(!$(".devtools-autocomplete-popup"), "Autocomplete list removed from DOM");
  }

  async function testKeyEventsWithAutocomplete() {
    // Clear the initial input
    $(".devtools-searchinput").focus();

    // ArrowDown
    synthesizeKey("VK_DOWN", {});
    await forceRender(component);
    is(refs.autocomplete.state.selectedIndex, 0, "selectedIndex is 0");
    ok($(".devtools-autocomplete-listbox .autocomplete-item:nth-child(1)")
      .className.includes("autocomplete-selected"),
      "Selection class applied");

    // ArrowUp should roll back to the bottom of the list
    synthesizeKey("VK_UP", {});
    await forceRender(component);
    is(refs.autocomplete.state.selectedIndex, 6, "ArrowUp works");

    // PageDown should take -5 places up
    synthesizeKey("VK_PAGE_UP", {});
    await forceRender(component);
    is(refs.autocomplete.state.selectedIndex, 1, "PageUp works");

    // PageDown should take +5 places down
    synthesizeKey("VK_PAGE_DOWN", {});
    await forceRender(component);
    is(refs.autocomplete.state.selectedIndex, 6, "PageDown works");

    // Home should take to the top of the list
    synthesizeKey("VK_HOME", {});
    await forceRender(component);
    is(refs.autocomplete.state.selectedIndex, 0, "Home works");

    // End should take to the bottom of the list
    synthesizeKey("VK_END", {});
    await forceRender(component);
    is(refs.autocomplete.state.selectedIndex, 6, "End works");

    // Key down in existing state should rollover to the top
    synthesizeKey("VK_DOWN", {});
    await forceRender(component);
    // Tab should select the component and hide popup
    synthesizeKey("VK_TAB", {});
    await forceRender(component);
    is(component.state.value, "ABC", "Tab hit selects the item");
    ok(!$(".devtools-autocomplete-popup"), "Tab hit hides the popup");

    // Activate popup by removing a key
    synthesizeKey("VK_BACK_SPACE", {});
    await forceRender(component);
    ok($(".devtools-autocomplete-popup"), "Popup is up");
    compareAutocompleteList($(".devtools-autocomplete-listbox"), ["ABC", "abc"]);

    // Enter key selection
    synthesizeKey("VK_UP", {});
    await forceRender(component);
    synthesizeKey("VK_RETURN", {});
    is(component.state.value, "abc", "Enter selection");
    ok(!$(".devtools-autocomplete-popup"), "Enter/Return hides the popup");

    // Escape should remove the autocomplete component
    synthesizeKey("VK_BACK_SPACE", {});
    await forceRender(component);
    synthesizeKey("VK_ESCAPE", {});
    await forceRender(component);
    ok(!$(".devtools-autocomplete-popup"),
      "Autocomplete list removed from DOM on Escape");
  }

  async function testMouseEventsWithAutocomplete() {
    $(".devtools-searchinput").focus();
    await setState(component, {
      value: "",
      focused: true,
    });
    await forceRender(component);

    // ArrowDown
    synthesizeKey("VK_DOWN", {});
    await forceRender(component);
    synthesizeMouseAtCenter($(".devtools-searchinput"), {}, window);
    await forceRender(component);
    is(component.state.focused, true, "Component should now be focused");

    sendString("pq");
    await forceRender(component);
    synthesizeMouseAtCenter(
      $(".devtools-autocomplete-listbox .autocomplete-item:nth-child(1)"),
      {}, window
    );
    await forceRender(component);
    is(component.state.value, "pqr", "Mouse click selects the item.");
    ok(!$(".devtools-autocomplete-popup"), "Mouse click on item hides the popup");
  }

  async function testTokenizedAutocomplete() {
    // Test for string "pqr ab" which should show list of ABC, abc
    sendString(" ab");
    await forceRender(component);
    compareAutocompleteList($(".devtools-autocomplete-listbox"), ["ABC", "abc"]);

    // Select the first element, value now should be "pqr ABC"
    synthesizeMouseAtCenter(
      $(".devtools-autocomplete-listbox .autocomplete-item:nth-child(1)"),
      {}, window
    );
    is(component.state.value, "pqr ABC", "Post Tokenization value selection");
  }

  add_task(async function () {
    await testSearchBoxWithAutocomplete();
    await testKeyEventsWithAutocomplete();
    await testMouseEventsWithAutocomplete();
    await testTokenizedAutocomplete();
  });
};
</script>
</body>
</html>
